+++ src/quacc/calculators/qchem/qchem.py
@@ -232,9 +232,6 @@ class QChem(FileIOCalculator):
         # Instantiate previous orbital coefficients
         self.prev_orbital_coeffs = None
 
-        if "directory" in self.fileiocalculator_kwargs:
-            raise NotImplementedError("The directory kwarg is not supported.")
-
         # Clean up parameters
         cleanup_attrs(self)
 
@@ -277,10 +274,13 @@ class QChem(FileIOCalculator):
 
         qc_input = make_qc_input(self, atoms)
 
-        write_qchem(qc_input, prev_orbital_coeffs=self.prev_orbital_coeffs)
+        write_qchem(
+            qc_input,
+            directory=self.directory,
+            prev_orbital_coeffs=self.prev_orbital_coeffs,
+        )
 
-    @staticmethod
-    def execute() -> int:
+    def execute(self) -> int:
         """
         Execute Q-Chem.
 
@@ -290,7 +290,7 @@ class QChem(FileIOCalculator):
             The return code.
         """
 
-        run_custodian()
+        run_custodian(directory=self.directory)
         return 0
 
     def read_results(self) -> None:
@@ -302,7 +302,7 @@ class QChem(FileIOCalculator):
         -------
         None
         """
-        results, prev_orbital_coeffs = read_qchem()
+        results, prev_orbital_coeffs = read_qchem(directory=self.directory)
         self.results = results
         self.prev_orbital_coeffs = prev_orbital_coeffs