+++ src/quacc/recipes/vasp/mp.py
@@ -115,6 +115,7 @@ def mp_gga_relax_job(
         atoms,
         calc_defaults=calc_defaults,
         calc_swaps=calc_kwargs,
+        report_mp_corrections=True,
         additional_fields={"name": "MP GGA Relax"},
         copy_files=copy_files,
     )
@@ -164,6 +165,7 @@ def mp_gga_static_job(
         atoms,
         calc_defaults=calc_defaults,
         calc_swaps=calc_kwargs,
+        report_mp_corrections=True,
         additional_fields={"name": "MP GGA Static"},
         copy_files=copy_files,
     )
@@ -220,6 +222,7 @@ def mp_metagga_prerelax_job(
         atoms,
         calc_defaults=calc_defaults,
         calc_swaps=calc_kwargs,
+        report_mp_corrections=True,
         additional_fields={"name": "MP Meta-GGA Pre-Relax"},
         copy_files=copy_files,
     )
@@ -271,6 +274,7 @@ def mp_metagga_relax_job(
         atoms,
         calc_defaults=calc_defaults,
         calc_swaps=calc_kwargs,
+        report_mp_corrections=True,
         additional_fields={"name": "MP Meta-GGA Relax"},
         copy_files=copy_files,
     )
@@ -324,6 +328,7 @@ def mp_metagga_static_job(
         atoms,
         calc_defaults=calc_defaults,
         calc_swaps=calc_kwargs,
+        report_mp_corrections=True,
         additional_fields={"name": "MP Meta-GGA Static"},
         copy_files=copy_files,
     )
@@ -441,15 +446,17 @@ def mp_metagga_relax_flow(
     MPMetaGGARelaxFlowSchema
         Dictionary of results. See the type-hint for the data structure.
     """
-    (
-        mp_metagga_prerelax_job_,
-        mp_metagga_relax_job_,
-        mp_metagga_static_job_,
-    ) = customize_funcs(
-        ["mp_metagga_prerelax_job", "mp_metagga_relax_job", "mp_metagga_static_job"],
-        [mp_metagga_prerelax_job, mp_metagga_relax_job, mp_metagga_static_job],
-        parameters=job_params,
-        decorators=job_decorators,
+    (mp_metagga_prerelax_job_, mp_metagga_relax_job_, mp_metagga_static_job_) = (
+        customize_funcs(
+            [
+                "mp_metagga_prerelax_job",
+                "mp_metagga_relax_job",
+                "mp_metagga_static_job",
+            ],
+            [mp_metagga_prerelax_job, mp_metagga_relax_job, mp_metagga_static_job],
+            parameters=job_params,
+            decorators=job_decorators,
+        )
     )
 
     # Run the prerelax