+++ src/quacc/runners/ase.py
@@ -135,7 +135,7 @@ def run_opt(
     max_steps: int = 1000,
     optimizer: Optimizer = BFGS,
     optimizer_kwargs: OptimizerKwargs | None = None,
-    store_intermediate_files: bool = False,
+    store_intermediate_results: bool = False,
     run_kwargs: dict[str, Any] | None = None,
     copy_files: SourceDirectory | dict[SourceDirectory, Filenames] | None = None,
 ) -> Optimizer:
@@ -163,7 +163,7 @@ def run_opt(
         Dictionary of kwargs for the optimizer. Takes all valid kwargs for ASE
         Optimizer classes. Refer to `_set_sella_kwargs` for Sella-related
         kwargs and how they are set.
-    store_intermediate_files
+    store_intermediate_results
         Whether to store the files generated at each intermediate step in the
         optimization. If enabled, they will be stored in a directory named
         `stepN` where `N` is the step number, starting at 0.
@@ -218,7 +218,7 @@ def run_opt(
 
     # Run optimization
     with traj, optimizer(atoms, **optimizer_kwargs) as dyn:
-        if store_intermediate_files:
+        if store_intermediate_results:
             opt = dyn.irun(fmax=fmax, steps=max_steps, **run_kwargs)
             for i, _ in enumerate(opt):
                 _copy_intermediate_files(