+++ tests/core/recipes/gulp_recipes/conftest.py
@@ -1,13 +1,8 @@
 import os
 from pathlib import Path
-from shutil import copy
 from unittest import mock
 
 import pytest
-from ase.calculators.calculator import FileIOCalculator
-from ase.calculators.gulp import GULP
-from ase.calculators.lj import LennardJones
-from ase.io import read
 
 FILE_DIR = Path(__file__).parent
 GULP_DIR = Path(FILE_DIR, "gulp_run")
@@ -20,18 +15,25 @@ def mock_settings_env_vars():
 
 
 def mock_execute(self, *args, **kwargs):
-    copy(GULP_DIR / "gulp.got.gz", "gulp.got.gz")
-    copy(GULP_DIR / "gulp.cif.gz", "gulp.cif.gz")
-    copy(GULP_DIR / "gulp.xyz.gz", "gulp.xyz.gz")
+    from shutil import copy
+
+    copy(GULP_DIR / "gulp.got.gz", Path(self.directory, "gulp.got.gz"))
+    copy(GULP_DIR / "gulp.cif.gz", Path(self.directory, "gulp.cif.gz"))
+    copy(GULP_DIR / "gulp.xyz.gz", Path(self.directory, "gulp.xyz.gz"))
 
 
 @pytest.fixture(autouse=True)
 def patch_execute(monkeypatch):
+    from ase.calculators.calculator import FileIOCalculator
+
     monkeypatch.setattr(FileIOCalculator, "execute", mock_execute)
 
 
 def mock_read_results(self, *args, **kwargs):
-    atoms = read("gulp.xyz.gz")
+    from ase.calculators.lj import LennardJones
+    from ase.io import read
+
+    atoms = read(Path(self.directory, "gulp.xyz.gz"))
     atoms.calc = LennardJones()
     atoms.get_potential_energy()
     self.calc = atoms.calc
@@ -40,6 +42,8 @@ def mock_read_results(self, *args, **kwargs):
 
 @pytest.fixture(autouse=True)
 def patch_read_results(monkeypatch):
+    from ase.calculators.gulp import GULP
+
     monkeypatch.setattr(GULP, "read_results", mock_read_results)
 
 
@@ -54,6 +58,8 @@ def mock_get_opt_state(self, **kwargs):
 
 @pytest.fixture(autouse=True)
 def patch_get_opt_state(monkeypatch):
+    from ase.calculators.gulp import GULP
+
     # Monkeypatch the .get_opt_state() method of the GULP calculator object so
     # we aren't fetching the actual state
     monkeypatch.setattr(GULP, "get_opt_state", mock_get_opt_state)