+++ tests/core/recipes/vasp_recipes/mocked/test_vasp_recipes.py
@@ -1,8 +1,18 @@
+from pathlib import Path
+from shutil import copy
+
+import numpy as np
 import pytest
 from ase.build import bulk, molecule
 
 from quacc import SETTINGS
-from quacc.recipes.vasp.core import double_relax_flow, relax_job, static_job
+from quacc.recipes.vasp.core import (
+    double_relax_flow,
+    non_scf_job,
+    ase_relax_job,
+    relax_job,
+    static_job,
+)
 from quacc.recipes.vasp.mp import (
     mp_gga_relax_flow,
     mp_gga_relax_job,
@@ -19,6 +29,9 @@ from quacc.recipes.vasp.slabs import slab_to_ads_flow
 from quacc.recipes.vasp.slabs import static_job as slab_static_job
 
 DEFAULT_SETTINGS = SETTINGS.model_copy()
+
+FILE_DIR = Path(__file__).parent
+MOCKED_DIR = FILE_DIR / "mocked_vasp_run"
 
 
 def test_static_job(tmp_path, monkeypatch):
@@ -147,6 +160,131 @@ def test_doublerelax_flow(tmp_path, monkeypatch):
     assert output["relax2"]["parameters"]["isif"] == 2
 
     assert double_relax_flow(atoms, relax1_kwargs={"kpts": [1, 1, 1]})
+
+
+def test_ase_relax_job(tmp_path, monkeypatch):
+    monkeypatch.chdir(tmp_path)
+
+    atoms = bulk("Al")
+
+    output = ase_relax_job(atoms)
+    assert output["nsites"] == len(atoms)
+    assert output["parameters"]["nsw"] == 0
+    assert output["parameters"]["lwave"] is False
+    assert output["parameters"]["lcharg"] is False
+    assert output["parameters"]["encut"] == 520
+    assert output["fmax"] == 0.01
+    assert len(output["trajectory_results"]) > 1
+
+
+def test_ase_relax_job2(tmp_path, monkeypatch):
+    monkeypatch.chdir(tmp_path)
+
+    atoms = bulk("Al")
+
+    output = ase_relax_job(atoms, opt_params={"store_intermediate_results": True})
+    assert output["nsites"] == len(atoms)
+    assert output["parameters"]["nsw"] == 0
+    assert output["parameters"]["lwave"] is False
+    assert output["parameters"]["lcharg"] is False
+    assert output["parameters"]["encut"] == 520
+    assert output["fmax"] == 0.01
+    assert len(output["trajectory_results"]) > 1
+    assert len(output["steps"]) == len(output["trajectory_results"])
+
+
+def test_non_scf_job1(tmp_path, monkeypatch):
+    monkeypatch.chdir(tmp_path)
+    copy(MOCKED_DIR / "vasprun.xml.gz", tmp_path / "vasprun.xml.gz")
+
+    atoms = bulk("Al")
+
+    output = non_scf_job(atoms, tmp_path)
+
+    assert "nsites" in output
+    assert "parameters" in output
+    assert "results" in output
+
+    assert output["parameters"]["lorbit"] == 11
+    assert output["parameters"]["lwave"] is False
+    assert output["parameters"]["lcharg"] is False
+    assert output["parameters"]["nsw"] == 0
+    assert output["parameters"]["isym"] == 2
+    assert output["parameters"]["icharg"] == 11
+    assert output["parameters"].get("kspacing") is None
+    assert output["parameters"]["nedos"] == 6001
+    assert output["parameters"]["kpts"] == [11, 11, 11]
+    assert output["parameters"]["ismear"] == -5
+    assert output["parameters"]["nbands"] == 99
+
+
+def test_non_scf_job2(tmp_path, monkeypatch, caplog):
+    monkeypatch.chdir(tmp_path)
+    copy(MOCKED_DIR / "vasprun.xml.gz", tmp_path / "vasprun.xml.gz")
+
+    atoms = bulk("Al")
+
+    output = non_scf_job(
+        atoms, tmp_path, preset="BulkSet", nbands_factor=1, calculate_optics=True
+    )
+
+    assert "nsites" in output
+    assert "parameters" in output
+    assert "results" in output
+
+    assert output["parameters"]["loptics"] is True
+    assert output["parameters"]["lreal"] is False
+    assert output["parameters"]["cshift"] == 1e-5
+    assert output["parameters"]["nbands"] == 82
+    assert output["parameters"]["ismear"] == -5
+    assert output["parameters"]["lorbit"] == 11
+    assert output["parameters"]["lwave"] is False
+    assert output["parameters"]["lcharg"] is False
+    assert output["parameters"]["nsw"] == 0
+    assert output["parameters"]["isym"] == 2
+    assert output["parameters"]["icharg"] == 11
+    assert output["parameters"].get("kspacing") is None
+
+
+@pytest.mark.parametrize("_is_metal", [True, False])
+def test_non_scf_job3(tmp_path, monkeypatch, _is_metal):
+    monkeypatch.chdir(tmp_path)
+
+    class DummyBandStructure:
+        def __init__(self):
+            pass
+
+        @staticmethod
+        def is_metal():
+            return _is_metal
+
+    monkeypatch.setattr(
+        "pymatgen.io.vasp.Vasprun.get_band_structure", DummyBandStructure
+    )
+    copy(MOCKED_DIR / "vasprun.xml.gz", tmp_path / "vasprun.xml.gz")
+
+    atoms = bulk("Al")
+
+    output = non_scf_job(atoms, tmp_path, preset="BulkSet", kpts_mode="line")
+    assert np.shape(output["parameters"]["kpts"]) == (250, 3)
+    if _is_metal:
+        assert output["parameters"]["sigma"] == 0.2
+        assert output["parameters"]["ismear"] == 1
+    else:
+        assert output["parameters"]["sigma"] == 0.01
+        assert output["parameters"]["ismear"] == 0
+
+
+def test_non_scf_job4(tmp_path, monkeypatch):
+    monkeypatch.chdir(tmp_path)
+
+    copy(MOCKED_DIR / "vasprun.xml.gz", tmp_path / "vasprun.xml.gz")
+    atoms = bulk("Al")
+
+    with pytest.raises(
+        ValueError, match="Supported kpoint modes are 'uniform' and 'line' at present"
+    ):
+        non_scf_job(atoms, tmp_path, kpts_mode="dummy")
 
 
 def test_slab_static_job(tmp_path, monkeypatch):
@@ -525,6 +663,12 @@ def test_mp_metagga_relax_flow(tmp_path, monkeypatch):
 
     output = mp_metagga_relax_flow(atoms)
     assert output["static"]["nsites"] == len(atoms)
+    assert output["prerelax"]["parameters"]["gga"] == "ps"
+    assert output["prerelax"]["parameters"]["ismear"] == 0
+    assert output["prerelax"]["parameters"]["pp"] == "pbe"
+    assert output["prerelax"]["parameters"]["magmom"] == [0.6]
+    assert output["relax1"]["parameters"]["magmom"] == [0.0]
+    assert output["relax2"]["parameters"]["magmom"] == [0.0]
     assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
     assert output["relax2"]["parameters"]["ediffg"] == -0.02
     assert output["relax2"]["parameters"]["encut"] == 680
@@ -532,13 +676,15 @@ def test_mp_metagga_relax_flow(tmp_path, monkeypatch):
     assert output["relax2"]["parameters"]["sigma"] == 0.05
     assert output["relax2"]["parameters"]["kspacing"] == 0.22
     assert output["relax2"]["parameters"]["pp"] == "pbe"
-    assert output["prerelax"]["parameters"]["gga"] == "ps"
-    assert output["prerelax"]["parameters"]["ismear"] == 0
-    assert output["prerelax"]["parameters"]["pp"] == "pbe"
 
     atoms = bulk("C")
+    atoms.set_initial_magnetic_moments([0.0, 0.0])
     output = mp_metagga_relax_flow(atoms)
-    assert output["static"]["nsites"] == len(atoms)
+    assert output["prerelax"]["parameters"]["ismear"] == 0
+    assert output["prerelax"]["parameters"]["pp"] == "pbe"
+    assert output["prerelax"]["parameters"]["magmom"] == [0.0, 0.0]
+    assert output["relax1"]["parameters"]["magmom"] == [0.0, 0.0]
+    assert output["relax2"]["parameters"]["magmom"] == [0.0, 0.0]
     assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
     assert output["relax2"]["parameters"]["ediffg"] == -0.02
     assert output["relax2"]["parameters"]["encut"] == 680
@@ -547,32 +693,37 @@ def test_mp_metagga_relax_flow(tmp_path, monkeypatch):
         0.28329488761304206
     )
     assert output["relax2"]["parameters"]["pp"] == "pbe"
-    assert output["prerelax"]["parameters"]["ismear"] == 0
-    assert output["prerelax"]["parameters"]["pp"] == "pbe"
+    assert output["static"]["nsites"] == len(atoms)
 
     atoms = molecule("O2")
+    atoms.set_initial_magnetic_moments([1.0, 0.0])
     atoms.center(vacuum=10)
     atoms.pbc = True
     output = mp_metagga_relax_flow(atoms)
+    assert output["prerelax"]["parameters"]["ismear"] == 0
+    assert output["prerelax"]["parameters"]["pp"] == "pbe"
+    assert output["prerelax"]["parameters"]["magmom"] == [1.0, 0.0]
+    assert output["relax1"]["parameters"]["magmom"] == [0.0, 0.0]
+    assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
+    assert output["relax2"]["parameters"]["ediffg"] == -0.02
+    assert output["relax2"]["parameters"]["encut"] == 680
+    assert output["relax2"]["parameters"]["ismear"] == 0
+    assert output["relax2"]["parameters"]["kspacing"] == pytest.approx(
+        0.28329488761304206
+    )
+    assert output["relax2"]["parameters"]["pp"] == "pbe"
+    assert output["relax2"]["parameters"]["magmom"] == [0.0, 0.0]
     assert output["static"]["nsites"] == len(atoms)
     assert output["static"]["parameters"]["ismear"] == -5
     assert output["static"]["parameters"]["nsw"] == 0
     assert output["static"]["parameters"]["algo"] == "fast"
-    assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
-    assert output["relax2"]["parameters"]["ediffg"] == -0.02
-    assert output["relax2"]["parameters"]["encut"] == 680
-    assert output["relax2"]["parameters"]["ismear"] == 0
-    assert output["relax2"]["parameters"]["kspacing"] == pytest.approx(
-        0.28329488761304206
-    )
-    assert output["relax2"]["parameters"]["pp"] == "pbe"
-    assert output["prerelax"]["parameters"]["ismear"] == 0
-    assert output["prerelax"]["parameters"]["pp"] == "pbe"
+    assert output["static"]["parameters"]["magmom"] == [0.0, 0.0]
 
 
 def test_mp_gga_relax_job():
     atoms = bulk("Ni") * (2, 1, 1)
     atoms[0].symbol = "O"
+    del atoms.arrays["initial_magmoms"]
     output = mp_gga_relax_job(atoms)
 
     assert output["nsites"] == len(atoms)
@@ -598,7 +749,7 @@ def test_mp_gga_relax_job():
         "lorbit": 11,
         "lreal": "auto",
         "lwave": False,
-        "magmom": [0.6, 0.6],
+        "magmom": [0.6, 5.0],
         "nelm": 100,
         "nsw": 99,
         "prec": "accurate",
@@ -612,6 +763,7 @@ def test_mp_gga_relax_job():
 def test_mp_gga_static_job():
     atoms = bulk("Ni") * (2, 1, 1)
     atoms[0].symbol = "O"
+    del atoms.arrays["initial_magmoms"]
     output = mp_gga_static_job(atoms)
     assert output["nsites"] == len(atoms)
     assert output["parameters"] == {
@@ -635,7 +787,7 @@ def test_mp_gga_static_job():
         "lorbit": 11,
         "lreal": False,
         "lwave": True,  # modified by us (sensible)
-        "magmom": [0.6, 0.6],
+        "magmom": [0.6, 5.0],
         "nelm": 100,
         "nsw": 0,
         "prec": "accurate",
@@ -648,8 +800,8 @@ def test_mp_gga_static_job():
 def test_mp_gga_relax_flow():
     atoms = bulk("Ni") * (2, 1, 1)
     atoms[0].symbol = "O"
+    del atoms.arrays["initial_magmoms"]
     output = mp_gga_relax_flow(atoms)
-    assert output["static"]["nsites"] == len(atoms)
     relax_params = {
         "algo": "fast",
         "ediff": 0.0001,
@@ -672,7 +824,7 @@ def test_mp_gga_relax_flow():
         "lorbit": 11,
         "lreal": "auto",
         "lwave": False,
-        "magmom": [0.6, 0.6],
+        "magmom": [0.6, 5.0],
         "nelm": 100,
         "nsw": 99,
         "prec": "accurate",
@@ -680,8 +832,11 @@ def test_mp_gga_relax_flow():
         "pp": "pbe",
         "setups": {"O": "", "Ni": "_pv"},
     }
+    relax2_params = relax_params.copy()
+    relax2_params["magmom"] = [0.0, 0.0]
+
     assert output["relax1"]["parameters"] == relax_params
-    assert output["relax2"]["parameters"] == relax_params
+    assert output["relax2"]["parameters"] == relax2_params
     assert output["static"]["parameters"] == {
         "algo": "fast",
         "ediff": 0.0001,
@@ -703,7 +858,7 @@ def test_mp_gga_relax_flow():
         "lorbit": 11,
         "lreal": False,
         "lwave": True,  # modified by us (sensible)
-        "magmom": [0.6, 0.6],
+        "magmom": [0.0, 0.0],
         "nelm": 100,
         "nsw": 0,
         "prec": "accurate",
@@ -716,6 +871,7 @@ def test_mp_gga_relax_flow():
 def test_mp_relax_flow_custom():
     atoms = bulk("Ni") * (2, 1, 1)
     atoms[0].symbol = "O"
+    del atoms.arrays["initial_magmoms"]
     output = mp_metagga_relax_flow(
         mp_gga_relax_flow(atoms, job_params={"mp_gga_relax_job": {"nsw": 0}})["static"][
             "atoms"