+++ tests/parsl/conftest.py
@@ -1,4 +1,3 @@
-import os
 from pathlib import Path
 
 TEST_RESULTS_DIR = Path(__file__).parent / "_test_results"
@@ -12,6 +11,8 @@ except ImportError:
 
 
 def pytest_sessionstart():
+    import os
+
     if parsl:
         parsl.load()
     file_dir = Path(__file__).parent